% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsTie.R
\name{baseline}
\alias{baseline}
\title{baseline}
\usage{
baseline(param = NULL)
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the baseline in the REM model}
}
\description{
This function specified the input for the baseline effect in the \code{formula} argument for the function \code{\link{remulateTie}} or \code{\link{remulateActor}}.
}
