% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remulateDyad.R
\name{burn_in_adj_mat}
\alias{burn_in_adj_mat}
\title{Function to generate random events for the burn in period}
\usage{
burn_in_adj_mat(actors, burn_in, rs)
}
\description{
Function to generate random events for the burn in period
}
