% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/remulateDyad.R
\name{get.density}
\alias{get.density}
\title{get.density
Function to compute the global density of the network. Can be used to check for degeneracy of the simulated network}
\usage{
get.density(evls, actors)

get.density(evls, actors)
}
\description{
density = number of unique edges in the network / total number of possible edges between nodes
evls event list for the network i.e a matrix with columns (dyad,time)

density = number of unique edges in the network / total number of possible edges between nodes
evls event list for the network i.e a matrix with columns (dyad,time)
}
