% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsTie.R
\name{minimum}
\alias{minimum}
\title{minimum}
\usage{
minimum(param = NULL, variable, attr_actors, scaling = c("none", "std"))
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the statistic in the REM model}

\item{variable}{character vector specifies the name of the column with covariate value in attr_actors data.frame}

\item{attr_actors}{data.frame object with rows specifying values of attr_actors for an actor. First column must contain actor id, Second column time when covariate value changes (default zero if no change), Third column contains values for the attributes with column name corresponding to variable name}

\item{scaling}{specifies the method for scaling the statistic. \code{"none"} [default] gives raw value of the statistic at time t, \code{"std"} the statistic is standardized per time}
}
\description{
Dyadic covariate: minimum attribute value for dyad (i,j) is the smaller of the attribute values for actors i , j
}
\details{
This function specifies the input for the minimum effect in the \code{formula} argument for the function \code{\link{remulateTie}} or \code{\link{remulateActor}}. Not to be used independently
}
