% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remulateDyad.R
\name{parse_formula}
\alias{parse_formula}
\title{covariates- list of data frames for each each effect in formula, NULL if the effect is not exogenous and a data frame with columns (id,time,value) if exogenous}
\usage{
parse_formula(formula, pred = FALSE)
}
\description{
covariates- list of data frames for each each effect in formula, NULL if the effect is not exogenous and a data frame with columns (id,time,value) if exogenous
}
