% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_formula_Dyad}
\alias{parse_formula_Dyad}
\title{covariates- list of data frames for each each effect in formula, NULL if the effect is not exogenous and a data frame with columns (id,time,value) if exogenous}
\usage{
parse_formula_Dyad(formula)
}
\description{
covariates- list of data frames for each each effect in formula, NULL if the effect is not exogenous and a data frame with columns (id,time,value) if exogenous
}
