% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsTie.R
\name{receive}
\alias{receive}
\title{receive}
\usage{
receive(param = NULL, variable, attributes, scaling = c("raw", "std"))
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the statistic in the REM model}

\item{variable}{character vector specifies the name of the column with covariate value in attributes data.frame}

\item{attributes}{data.frame object with rows specifying values of attributes for an actor. First column must contain actor id, Second column time when covariate value changes (default zero if no change), Third column contains values for the attributes with column name corresponding to variable name}

\item{scaling}{specifies the method for scaling the statistic. \code{"raw"} [default] gives raw value of the statistic at time t, \code{"std"} the statistic is standardized per time}
}
\description{
Receiver covariate: The tendency to create an event i->j when j has a high attribute value.
}
\details{
This function specifies the input for the receive effect in the \code{formula} argument for the function \code{\link{remulateTie}} or \code{\link{remulateActor}}. Not to be used independently
}
