% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remulateDyad.R
\name{remulateDyad}
\alias{remulateDyad}
\title{Simulate REM Dyad}
\usage{
remulateDyad(
  formula,
  actors,
  M,
  burn_in = 0,
  risk_set = NULL,
  waiting_time = c("exp", "weibull", "gompertz"),
  time_param = 1,
  memory = c("full", "window", "window_m", "brandes", "vu"),
  memory_param = NULL,
  seeds = NULL
)
}
\arguments{
\item{actors}{Vector of actor IDs}

\item{M}{Number of events to generate}

\item{burn_in}{Number of random events to sample before beginning with the data generation}

\item{risk_set}{risk set, a dataframe wtih sender in 1st column and receiver in second column indicating which pair cannot be in risk set}

\item{memory}{[Optional] (default = full) String indicating which
 memory type to use. "full" uses the entire event history to compute statistics, "window" memory indicates a window in the past upto
which occured events will be remembered for computing statistics, "brandes" memory type uses past events
weighted by their time, "vu" memory type uses past events weighted by 1/time difference}

\item{memory_param}{[Optional] memory_param value > 0. For memory type "window" this parameter indicates the length (in time units) of the window, For memory type "brandes" the memory_param will be the half-life i.e the time until an event has a weight of one half.}

\item{form}{an object of class \code{"\link[stats]{formula}"}: a symbolic description of statistics used to generate the data.}
}
\value{
edgelist

statistics 3d cube of statistics of dimensions M x D x P

evls: event list a matrix with columns (dyad,time)

actors_map : mapping of actor IDs provided by user to the integer ids used in the edgelist

riskset: risket set used for the dyad position in the statistics and event list
}
\description{
A function to simulate relational event date by sampling from a
tie based relational event model.
}
\examples{
form <- ~baseline(1)+inertia(0.1)+reciprocity(0.4)
actors <- c(1:10)
remulateDyad(form,actors,100)
}
