% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsTie.R
\name{rrankSend}
\alias{rrankSend}
\title{rrankSend}
\usage{
rrankSend(param = NULL)
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the statistic in the REM model}
}
\description{
This function specifies the input for the rrankSend effect in the \code{formula} argument for the function \code{\link{remulateTie}}. Not to be used independently
}
