% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsTie.R
\name{send}
\alias{send}
\title{send}
\usage{
send(param = NULL, variable, attributes, scaling = c("raw", "std"))
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the statistic in the REM model}

\item{variable}{character vector specifies the name of the column with covariate value in data dataframe}

\item{scaling}{specifies the method for scaling the statistic. \code{"raw"} [default] gives raw value of the statistic at time t, \code{"std"} the statistic is standardized per time}

\item{data}{data.frame object with rows specifying values of data for an actor. First column must contain actor id, Second column time when covariate value changes (default zero if no change), Third column contains values for the data with column name corresponding to variable name}
}
\description{
Sender covariate: The tendency to create an event i->j when i has a high attribute value.
}
