% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsDyad.R
\name{difference}
\alias{difference}
\title{difference}
\usage{
difference(param = NULL, variable, covariates, scaling = c("raw", "std"))
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the statistic in the REM model}

\item{variable}{character vector: specifies the name of the column with covariate value in covariates dataframe}

\item{covariates}{data.frame object with rows specifying values of covariates for an actor. First column must contain actor id, Second column time when covariate value changes (default zero if no change), Third column contains values for the covariates with column name corresponding to variable name}

\item{scaling}{specifies the method for scaling the statistic. \code{"raw"} [default] gives raw value of the statistic at time t, \code{"std"} the statistic is standardized per time}
}
\description{
difference
}
