% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsDyad.R
\name{interact}
\alias{interact}
\title{interact}
\usage{
interact(param = NULL, scaling = c("raw", "std"))
}
\arguments{
\item{param}{numeric value or function with time parameter. Specifies the value of the effect for the statistic in the REM model}

\item{scaling}{the method for scaling the interaction statistic. \code{"raw"} [default] gives raw value of the statistic at time t, \code{"std"} the statistic is standardized per time point after multiplication of the interaction components}
}
\description{
This function specifies the input for the interact effect in the \code{formula} argument for the function \code{\link{remulateDyad}}. The interaction is always between the first two terms of the formula argument. Not to be used independently
}
