% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remulate-package.R
\docType{package}
\name{remulate-package}
\alias{remulate}
\alias{remulate-package}
\title{remulate: Remulate - Simulate Dynamic Networks from Relational Event Models}
\description{
This package allows for model based generation of relational event networks under tie and actor oriented models while providing flexibility to model temporal variability, heterogeneity, changes in the network composition and to account for memory decay in dynamic social networks. The development of this package was supported by a Vidi Grant (452-17-006) awarded by the Netherlands Organization for Scientific Research (NWO) Grant and an ERC Starting Grant (758791).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/remulate}
}

}
\author{
\strong{Maintainer}: Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu}

Other contributors:
\itemize{
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu} [contributor]
  \item Giuseppe Arena \email{g.arena@tilburguniversity.edu} [contributor]
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu} [contributor]
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu} [contributor]
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu} [contributor]
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
