% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remulate-package.R
\docType{package}
\name{remulate-package}
\alias{remulate}
\alias{remulate-package}
\title{remulate: Simulate Dynamic Networks from Relational Event Models}
\description{
Model based simulation of dynamic networks under tie-oriented (Butts, C., 2008, \doi{10.1111/j.1467-9531.2008.00203.x}) and actor-oriented (Stadtfeld, C., & Block, P., 2017, \doi{10.15195/v4.a14}) relational event models. Supports simulation from a variety of relational event model extensions, including temporal variability in effects, heterogeneity through dyadic latent class relational event models (DLC-REM), random effects, blockmodels, and memory decay in relational event models (Lakdawala, R., 2024 \doi{10.48550/arXiv.2403.19329}). The development of this package was supported by a Vidi Grant (452-17-006) awarded by the Netherlands Organization for Scientific Research (NWO) Grant and an ERC Starting Grant (758791).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/remulate}
}

}
\author{
\strong{Maintainer}: Rumana Lakdawala \email{rumanalakdawala@gmail.com} (\href{https://orcid.org/0000-0002-9992-6035}{ORCID})

Other contributors:
\itemize{
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu} [contributor]
  \item Giuseppe Arena \email{g.arena@tilburguniversity.edu} [contributor]
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu} [contributor]
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu} [contributor]
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu} [contributor]
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
